Add files via upload
In this Python project, I have worked with the United States Social Security Administration (SSA)
which has made data on the frequency of baby names from 1880 to 2020 available. The project
requires analyzing the data considering different methods. I have visualized the most frequent names,
determine the naming trends, and come up with the most popular names for a certain year.